adding rating from csv option

+++ app/models/movie.rb
@@ -4,6 +4,6 @@ class Movie < ApplicationRecord
   validates :title, presence: true, uniqueness: { scope: :director }
 
   def average_score
-    user_movies.average(:score).to_f
+    user_movies.average(:score).to_i
   end
 end
